+++ src/__fixtures__/destructured-properties/invalid-comments.ts
