+++ src/__fixtures__/destructured-properties/valid-comments.ts
